backend/pyproject.toml
[project]
name = "yuki-backend"
version = "0.1.0"
description = "API FastAPI pour l’assistant Yuki"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.17",
    "httpx>=0.28.0",
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "pillow>=11.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "psycopg[binary]>=3.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
